requirements.txt
streamlit==1.37.0
tensorflow==2.18.0
numpy
pillow